// pkgs/akanjs/signal/types.ts
interface SerializedSignalOption {
  args: SerializedArg[];
  path?: string;
  prefix?: false | string;
  globalPrefix?: false;
  guards?: string[];
}
export interface SerializedSlice extends SerializedSignalOption {}

export interface SerializedReturns {
  refName: Exclude<DefaultPrimitiveName, "Map" | "Upload"> | (string & {});
  modelType?: "input" | "full" | "light" | "insight" | "scalar"; // undefined when primitive
  arrDepth?: number;
  partial?: string[];
  nullable?: boolean;
}
export interface SerializedArg {
  type: ArgType;
  refName: Exclude<DefaultPrimitiveName, "Map"> | (string & {});
  name: string;
  modelType?: "input" | "object" | "insight" | "scalar"; // undefined when primitive
  arrDepth?: number;
  nullable?: boolean;
  example?: string | number | boolean | Date;
  enum?: string;
}
export interface SerializedEndpoint extends SerializedSignalOption {
  type: "query" | "mutation" | "pubsub" | "message";
  returns: SerializedReturns;
}
export interface SerializedFilter {
  filter: { [key: string]: SerializedArg[] };
  sortKeys: string[];
}

export interface SerializedSignal {
  prefix?: string;
  slice?: { [key: string]: SerializedSlice };
  endpoint: { [key: string]: SerializedEndpoint };
  filter?: SerializedFilter;
  getGuards?: string[];
  cruGuards?: string[];
}